Stats can now be regenerated as FLAT or PERCENTAGE based values. This includes the Health, Mana, Stamina, Stellium stats.
By default, the stats are regenerated off the flat value of their respective Regeneration stats. Example, if Health Regeneration is 10, the player will gain 10 HP upon regeneration.

For percentage, the stats will be regenerated with the following formula (Stat Regen * Max Stat). So for example, if your Health Regeneration is .5 and your Max Health is 100, you'll gain 50 HP upon regeneration.

Refresh your config.yml or add this directly to it if you wish to adjust them from the default true.
```# Should these Stats regenrate with a FLAT value or PERCENTAGE value?
   # By default, they are set to TRUE, which means they will regenerate
   # based off the FLAT Regeneration value for the respective stat
   # Setting a stat to FALSE means that it will regenerate
   # with the formula (Stat Regen * Max Stat) for example (Mana Regeneration * Max Mana)
   stat-regen-flat:
       health: true
       mana: true
       stamina: true
       stellium: true
```